fix radio Rx problem
The RX Fifo was not being drained which was preventing the packet
trigger from happening.

This was in part because I had turned off the auto-clear on CRC
error as part of debugging earlier, which would allow trash
to accumulate in the FIFO. Plus I needed to add some FIFO
drains just to make sure the FIFOs cleared out in case corrupt
trash appeared in it.