add a separate subdir for the heart badge source
This is based off of the last viable public commit from
heart-badge.

The heart-badge branch is now DEPRECATED, all changes for the
heart-badge should be made in the heart-badge subdir.

Notably, the "fx" directory is now a symlink to the "src" directory,
so any changes to effects should be reflected automagically
on the heart badge.

To build the heart badge simply cd to src-heart/ and run make.
The build will be in build/cubemaster*

Let's hope that git manages the "fx" symlinks correctly across
everyone else's repos -- this might require some plumbing,
especially for Jeanie because she's on a Windows machine which
may not support symlinks like Linux does